+++ play.it-2/games/play-torment-tides-of-numenera.sh
@@ -1,4 +1,4 @@
-#!/bin/sh -e
+#!/bin/sh
 set -o errexit
 
 ###
@@ -30,18 +30,16 @@ set -o errexit
 
 ###
 # Torment: Tides of Numenera
-# build native Linux packages from the original installers
+# build native packages from the original installers
 # send your bug reports to vv221@dotslashplay.it
 ###
 
-script_version=20180224.1
+script_version=20190724.2
 
 # Set game-specific variables
 
 GAME_ID='torment-tides-of-numenera'
 GAME_NAME='Torment: Tides of Numenera'
-
-ARCHIVES_LIST='ARCHIVE_GOG'
 
 ARCHIVE_GOG='gog_torment_tides_of_numenera_2.3.0.4.sh'
 ARCHIVE_GOG_URL='https://www.gog.com/game/torment_tides_of_numenera_legacy_edition_content'
@@ -49,34 +47,31 @@ ARCHIVE_GOG_MD5='839337b42a1618f3b445f363eca210d3'
 ARCHIVE_GOG_SIZE='9300000'
 ARCHIVE_GOG_VERSION='1.1.0-gog2.3.0.4'
 
-ARCHIVE_DOC_PATH='data/noarch/docs'
-ARCHIVE_DOC_FILES='./*'
+ARCHIVE_DOC_DATA_PATH='data/noarch/docs'
+ARCHIVE_DOC_DATA_FILES='*'
 
 ARCHIVE_GAME_BIN_PATH='data/noarch/game'
-ARCHIVE_GAME_BIN_FILES='./TidesOfNumenera ./TidesOfNumenera_Data/Mono/x86_64 ./TidesOfNumenera_Data/Plugins'
+ARCHIVE_GAME_BIN_FILES='TidesOfNumenera TidesOfNumenera_Data/Mono/x86_64 TidesOfNumenera_Data/Plugins'
 
 ARCHIVE_GAME_AUDIO_PATH='data/noarch/game'
-ARCHIVE_GAME_AUDIO_FILES='./TidesOfNumenera_Data/StreamingAssets/Audio'
+ARCHIVE_GAME_AUDIO_FILES='TidesOfNumenera_Data/StreamingAssets/Audio'
 
 ARCHIVE_GAME_RESOURCES_PATH='data/noarch/game'
-ARCHIVE_GAME_RESOURCES_FILES='./TidesOfNumenera_Data/resources.assets*'
+ARCHIVE_GAME_RESOURCES_FILES='TidesOfNumenera_Data/resources.assets*'
 
 ARCHIVE_GAME_DATA_PATH='data/noarch/game'
-ARCHIVE_GAME_DATA_FILES='./TidesOfNumenera_Data'
+ARCHIVE_GAME_DATA_FILES='TidesOfNumenera_Data'
 
 DATA_DIRS='./logs'
 
 APP_MAIN_TYPE='native'
-APP_MAIN_PRERUN='pulseaudio --start
-export LANG="en_US.UTF-8"'
+APP_MAIN_PRERUN='export LANG=C'
 APP_MAIN_EXE='TidesOfNumenera'
 # shellcheck disable=SC2016
 APP_MAIN_OPTIONS='-logFile ./logs/$(date +%F-%R).log'
-APP_MAIN_ICONS_LIST='APP_MAIN_ICON'
 APP_MAIN_ICON='TidesOfNumenera_Data/Resources/UnityPlayer.png'
-APP_MAIN_ICON_RES='128'
 
-PACKAGES_LIST='PKG_AUDIO PKG_RESOURCES PKG_DATA PKG_BIN'
+PACKAGES_LIST='PKG_BIN PKG_AUDIO PKG_RESOURCES PKG_DATA'
 
 PKG_AUDIO_ID="${GAME_ID}-audio"
 PKG_AUDIO_DESCRIPTION='audio'
@@ -88,56 +83,55 @@ PKG_DATA_ID="${GAME_ID}-data"
 PKG_DATA_DESCRIPTION='data'
 
 PKG_BIN_ARCH='64'
-PKG_BIN_DEPS_DEB="$PKG_AUDIO_ID, $PKG_RESOURCES_ID, $PKG_DATA_ID, libgl1-mesa-glx | libgl1, libsdl2-2.0-0, pulseaudio"
-PKG_BIN_DEPS_ARCH="$PKG_AUDIO_ID $PKG_RESOURCES_ID $PKG_DATA_ID libgl sdl2 pulseaudio"
+PKG_BIN_DEPS="$PKG_AUDIO_ID $PKG_RESOURCES_ID $PKG_DATA_ID glx sdl2"
 
 # Load common functions
 
-target_version='2.3'
+target_version='2.11'
 
 if [ -z "$PLAYIT_LIB2" ]; then
-	[ -n "$XDG_DATA_HOME" ] || XDG_DATA_HOME="$HOME/.local/share"
-	if [ -e "$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh" ]; then
-		PLAYIT_LIB2="$XDG_DATA_HOME/play.it/play.it-2/lib/libplayit2.sh"
-	elif [ -e './libplayit2.sh' ]; then
-		PLAYIT_LIB2='./libplayit2.sh'
-	else
-		printf '\n\033[1;31mError:\033[0m\n'
-		printf 'libplayit2.sh not found.\n'
-		exit 1
-	fi
+	: "${XDG_DATA_HOME:="$HOME/.local/share"}"
+	for path in\
+		"$PWD"\
+		"$XDG_DATA_HOME/play.it"\
+		'/usr/local/share/games/play.it'\
+		'/usr/local/share/play.it'\
+		'/usr/share/games/play.it'\
+		'/usr/share/play.it'
+	do
+		if [ -e "$path/libplayit2.sh" ]; then
+			PLAYIT_LIB2="$path/libplayit2.sh"
+			break
+		fi
+	done
 fi
-#shellcheck source=play.it-2/lib/libplayit2.sh
+if [ -z "$PLAYIT_LIB2" ]; then
+	printf '\n\033[1;31mError:\033[0m\n'
+	printf 'libplayit2.sh not found.\n'
+	exit 1
+fi
+# shellcheck source=play.it-2/lib/libplayit2.sh
 . "$PLAYIT_LIB2"
 
 # Extract game data
 
 extract_data_from "$SOURCE_ARCHIVE"
+prepare_package_layout
+rm --recursive "$PLAYIT_WORKDIR/gamedata"
 
-PKG='PKG_AUDIO'
-organize_data 'GAME_AUDIO' "$PATH_GAME"
-
-PKG='PKG_RESOURCES'
-organize_data 'GAME_RESOURCES' "$PATH_GAME"
+# Get icon
 
 PKG='PKG_DATA'
-organize_data 'DOC'       "$PATH_DOC"
-organize_data 'GAME_DATA' "$PATH_GAME"
-
-PKG='PKG_BIN'
-organize_data 'GAME_BIN' "$PATH_GAME"
-
-rm --recursive "$PLAYIT_WORKDIR/gamedata"
+icons_get_from_package 'APP_MAIN'
 
 # Write launchers
 
-write_launcher 'APP_MAIN'
+PKG='PKG_BIN'
+launchers_write 'APP_MAIN'
 
 # Build package
 
-postinst_icons_linking 'APP_MAIN'
-write_metadata 'PKG_DATA'
-write_metadata 'PKG_BIN' 'PKG_AUDIO' 'PKG_RESOURCES'
+write_metadata
 build_pkg
 
 # Clean up
